application/backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
numpy==1.24.0
pillow==11.1.0
tensorflow==2.13.0
joblib==1.4.2
scikit-learn==1.0.2
gunicorn==20.1.0
Werkzeug==2.2.2
